+++ Scripts/Enums/DSC_NetworkEventType.cs
@@ -1,0 +1,10 @@
+namespace DSC.Network
+{
+    public enum DSC_NetworkEventType
+    {
+        PreStartNetwork,
+        PostStartNetwork,
+        PreStopNetwork,
+        PostStopNetwork
+    }
+}