+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.0.1-preview.1] - 2021-03-30
+- Add DSC_NetworkEventType enum.
+- Add DSC_Network script.
+- Add DSC_Event_Network script.
+- Add DSC_Event_NetworkMode script.
+- Add DSC_Event_RunEventsByNetworkEvent script.
+- Add NetworkTemplate sample.
+
 ## [0.0.0-preview.5] - 2021-03-25
 - Add NetworkMode enum.
 - Add DSC_Network_HUD script.